requirements.txt
flask==3.0.2
google-cloud-storage==2.14.0
google-generativeai==0.7.2